langgraph-qa-bot/requirements.txt
langgraph
python-dotenv
langchain-openai